+++ com.unity.render-pipelines.high-definition/Editor/Material/UIBlocks/DecalSurfaceInputsUIBlock.cs
@@ -25,7 +25,7 @@ namespace UnityEditor.Rendering.HighDefinition
             public static GUIContent normalOpacityChannelText = new GUIContent("Normal Opacity Channel", "Specifies the source this Material uses as opacity for its Normal Map.");
             public static GUIContent smoothnessRemappingText = new GUIContent("Smoothness Remapping", "Controls a remap for the smoothness channel in the Mask Map.");
             public static GUIContent smoothnessText = new GUIContent("Smoothness", "Controls the smoothness of the decal.");
-            public static GUIContent metallicScaleText = new GUIContent("Metallic Scale", "Controls a scale factor for the metallic channel in the Mask Map.");
+            public static GUIContent metallicRemappingText = new GUIContent("Metallic Remapping", "Controls a remap for the metallic channel in the Mask Map.");
             public static GUIContent metallicText = new GUIContent("Metallic", "Controls the metallic of the decal.");
             public static GUIContent aoRemappingText = new GUIContent("Ambient Occlusion Remapping", "Controls a remap for the ambient occlusion channel in the Mask Map.");
             public static GUIContent aoText = new GUIContent("Ambient Occlusion", "Controls the ambient occlusion of the decal.");
@@ -95,6 +95,9 @@ namespace UnityEditor.Rendering.HighDefinition
         MaterialProperty smoothnessRemapMax = new MaterialProperty();
         const string kSmoothnessRemapMax = "_SmoothnessRemapMax";
 
+        MaterialProperty metallicRemapMin = new MaterialProperty();
+        const string kMetallicRemapMin = "_MetallicRemapMin";
+
         MaterialProperty metallicScale = new MaterialProperty();
         const string kMetallicScale = "_MetallicScale";
 
@@ -159,6 +162,7 @@ namespace UnityEditor.Rendering.HighDefinition
             AORemapMax = FindProperty(kAORemapMax);
             smoothnessRemapMin = FindProperty(kSmoothnessRemapMin);
             smoothnessRemapMax = FindProperty(kSmoothnessRemapMax);
+            metallicRemapMin = FindProperty(kMetallicRemapMin);
             metallicScale = FindProperty(kMetallicScale);
             AO = FindProperty(kAO);
             smoothness = FindProperty(kSmoothness);
@@ -225,7 +229,15 @@ namespace UnityEditor.Rendering.HighDefinition
             {
                 if (perChannelMask)
                 {
-                    materialEditor.ShaderProperty(metallicScale, Styles.metallicScaleText);
+                    float MetalRemapMinValue = metallicRemapMin.floatValue;
+                    float MetalRemapMaxValue = metallicScale.floatValue;
+                    EditorGUI.BeginChangeCheck();
+                    EditorGUILayout.MinMaxSlider(Styles.metallicRemappingText, ref MetalRemapMinValue, ref MetalRemapMaxValue, 0.0f, 1.0f);
+                    if (EditorGUI.EndChangeCheck())
+                    {
+                        metallicRemapMin.floatValue = MetalRemapMinValue;
+                        metallicScale.floatValue = MetalRemapMaxValue;
+                    }
 
                     float AORemapMinValue = AORemapMin.floatValue;
                     float AORemapMaxValue = AORemapMax.floatValue;